requirements.txt
# Core Data Processing
pandas==2.1.4
numpy>=1.26.0

# Web Framework & API
streamlit>=1.29.0
fastapi>=0.104.1
uvicorn>=0.24.0

# AI & Machine Learning
openai>=1.6.1

# Data Validation & Models
pydantic>=2.5.2

# String Matching & Processing
fuzzywuzzy>=0.18.0
python-levenshtein>=0.21.1

# Environment & Configuration
python-dotenv>=1.0.0

# Excel File Processing
openpyxl>=3.1.2
xlrd>=2.0.1

# HTTP Requests
requests>=2.31.0

# Type Hints (for older Python versions)
typing-extensions>=4.8.0

# Development & Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1

# Optional: For enhanced logging and debugging
colorama==0.4.6
